requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
python-linkedin-v2==0.9.4
requests==2.31.0
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-auth==2.40.3
openpyxl==3.1.5
reportlab==4.4.3
matplotlib==3.10.5
pandas==2.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9 